+++ internal/provider/schema_firewall_ruleset_attachment.go
@@ -9,6 +9,11 @@ func schemaFirewallRulesetAttachment() tfsdk.Schema {
 	return tfsdk.Schema{
 		Description: "Attach a firewall ruleset to inbound, outbound, and local traffic.",
 		Attributes: map[string]tfsdk.Attribute{
+			"id": {
+				Description: "The identifier of the resource. This will always be the interface. It is present only for legacy purposes.",
+				Type:        types.StringType,
+				Computed:    true,
+			},
 			"interface": {
 				Type:          types.StringType,
 				Required:      true,