+++ internal/provider/utils.go
@@ -1,0 +1,5 @@
+package provider 
+
+func strptr(s string) *string {
+	return &s
+}