+++ internal/utils/utils.go
@@ -1,5 +1,9 @@
 package utils
 
+type hasID interface {
+	GetID() string
+}
+
 func WithPrefix(prefix string, arr []string) []string {
 	tmp := make([]string, len(arr))
 	for i := 0; i < len(arr); i++ {